requirements.txt
langchain==0.1.6
langchain-community==0.0.20
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.12.0
pydantic==2.6.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
email-validator==2.1.0
sqlalchemy==2.0.25
pandas==2.2.0
python-dateutil==2.8.2
openpyxl==3.1.2
